Adjusted backend integration based on feedback
