+++ server/src/service/business/geoChecksService.ts
@@ -1,9 +1,9 @@
 import type { Monitor, GeoCheck } from "@/types/index.js";
 import type { GeoCheckResult } from "@/types/geoCheck.js";
+import type { IGlobalPingService } from "@/types/globalping.js";
 import { Types } from "mongoose";
 import type { IGeoChecksRepository } from "@/repositories/index.js";
 import type { IMonitorsRepository } from "@/repositories/index.js";
-import type { IGlobalPingService } from "@/service/infrastructure/globalPingService.js";
 import type { ILogger } from "@/utils/logger.js";
 import { AppError } from "@/utils/AppError.js";
